Attempted impl of large steps in inv. rendering
This is an attempt at an implementation for large steps in inverse rendering.
It doesn't seem to work, despite closely mirroring the code given in their code release. I
suspect this may be due to numerical errors in solving, or rasterization. I would continue work,
but I am unsure if I have time to continue on it in the near future.